1) Moved to few common code to utilities_common/common.py 2) Avoided kube.py importing main.py
